requirements.txt
pyserial>=3.5
odrive==0.5.1
